requirements.txt
pillow
psutil
torch
torchvision
torchaudio
nuitka
nuitka[python3]



